+++ workflows/ingestion/cellranger/test/integration_test.sh
@@ -10,7 +10,7 @@ export NXF_VER=21.10.6
 
 bin/nextflow \
   run . \
-  -main-script workflows/1_ingestion/bd_rhapsody_wta/main.nf \
+  -main-script workflows/ingestion/cellranger/main.nf \
   -entry test_wf \
   -with-docker \
   -resume