+++ CHANGELOG.md
@@ -6,7 +6,14 @@
 
 ## BUG FIXES
 
-* `workflows/utils/utils.nf`: Fix error message when required parameter is not specified.
+* `workflows/utils/viash_workflow_helper.nf`: Renamed `utils.nf` to `viash_workflow_helper.nf`.
+
+* `workflows/utils/viash_workflow_helper.nf`: Fix error message when required parameter is not specified.
+
+* `workflows/utils/viash_workflow_helper.nf`: Added helper functions:
+  - `readConfig`: Read a Viash config from a yaml file.
+  - `viashChannel`: Create a channel from the Viash config and the params object.
+  - `helpMessage`: Print a help message and exit.
 
 * `mapping/bd_rhapsody_wta`: Allow unit test to be run on less powerful systems (at least 4GB ram).