+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `convert/from_bdrhap_to_h5mu`: Merge one or more BD rhapsody outputs into an h5mu file.
 
+* `split/split`: Split the modalities from a single .h5mu multimodal sample into seperate .h5mu files. 
+
 ## BUG FIXES
 
 * `workflows/utils/viash_workflow_helper.nf`: Renamed `utils.nf` to `viash_workflow_helper.nf`.