+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * `convert/from_bdrhap_to_h5mu`: Merge one or more BD rhapsody outputs into an h5mu file.
 
+* `split/split_modalities`: Split the modalities from a single .h5mu multimodal sample into seperate .h5mu files. 
+
 * `integrate/concat`: Combine data from multiple samples together.
 
 ## BUG FIXES